RSPY-468: Updated tests configuration
